refactor(BadgeList): refactor to ts
