+++ packages/orbit-components/src/BadgeList/BadgeListItem/index.d.ts
@@ -1,11 +1,10 @@
 // Type definitions for @kiwicom/orbit-components
 // Project: http://github.com/kiwicom/orbit
-
 import * as React from "react";
 
 import * as Common from "../../common/common";
 
-type Type = "neutral" | "info" | "success" | "warning" | "critical";
+export type Type = "neutral" | "info" | "success" | "warning" | "critical";
 
 export interface Props extends Common.Globals {
   readonly children: React.ReactNode;
