--- packages/orbit-components/src/BadgeList/BadgeListItem/index.jsx.flow
@@ -1,13 +1,0 @@
-// @flow
-import * as React from "react";
-
-import type { Globals } from "../../common/common.js.flow";
-
-export type Props = {|
-  +children: React.Node,
-  +icon: React.Element<any>,
-  +type?: "neutral" | "info" | "success" | "warning" | "critical",
-  ...Globals,
-|};
-
-declare export default React.ComponentType<Props>;
