+++ packages/orbit-components/src/BadgeList/__tests__/index.test.tsx
@@ -1,4 +1,3 @@
-// @flow
 import * as React from "react";
 import { screen, render } from "@testing-library/react";
