--- packages/orbit-components/src/BadgeList/consts.js
@@ -1,8 +1,0 @@
-// @flow
-export default {
-  NEUTRAL: "neutral",
-  INFO: "info",
-  SUCCESS: "success",
-  WARNING: "warning",
-  CRITICAL: "critical",
-};
