+++ packages/orbit-components/src/BadgeList/consts.ts
@@ -1,0 +1,7 @@
+export enum TYPE_OPTIONS {
+  NEUTRAL = "neutral",
+  INFO = "info",
+  SUCCESS = "success",
+  WARNING = "warning",
+  CRITICAL = "critical",
+}
