--- packages/orbit-components/src/BadgeList/index.jsx.flow
@@ -1,17 +1,0 @@
-// @flow
-/*
-  DOCUMENTATION: https://orbit.kiwi/components/badgelist/
-*/
-import * as React from "react";
-
-import type { Globals } from "../common/common.js.flow";
-import type { Props as BadgeListItemProps } from "./BadgeListItem";
-
-export type Props = {|
-  +children: React.Node,
-  ...Globals,
-|};
-
-declare export var BadgeListItem: React.ComponentType<BadgeListItemProps>;
-
-declare export default React.ComponentType<Props>;
